Diseño: modificamos los colores, botones, todo la interfaz de los usuarios doctores
