+++ frontend/src/app/doctor-dashboard/components/doctor-bitacoras/doctor-bitacoras.component.css
@@ -1,0 +1,317 @@
+/* ===== DOCTOR BITÁCORAS - ESTILOS MEJORADOS ===== */
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&display=swap');
+
+.bitacoras-wrapper {
+    display: flex;
+    flex-direction: column;
+    gap: 2rem;
+}
+
+/* Encabezado */
+.section-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    flex-wrap: wrap;
+    gap: 1rem;
+    margin-bottom: 0.5rem;
+}
+
+.section-header h2 {
+    font-size: 1.8rem;
+    font-weight: 700;
+    background: linear-gradient(135deg, #E2E8F0, #A5F3C3);
+    background-clip: text;
+    -webkit-background-clip: text;
+    color: transparent;
+    margin: 0;
+}
+
+.section-header-actions {
+    display: flex;
+    gap: 0.8rem;
+}
+
+/* Botón CSV */
+.btn-csv {
+    background: #1565C0;
+    color: white;
+    border: none;
+    padding: 0.6rem 1.2rem;
+    border-radius: 40px;
+    font-weight: 600;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.btn-csv:hover:not(:disabled) {
+    background: #0D47A1;
+    transform: translateY(-2px);
+}
+
+.btn-csv:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+/* Filtros */
+.filtros-bar {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: center;
+    gap: 0.8rem;
+    background: #1E293B;
+    padding: 1rem 1.5rem;
+    border-radius: 60px;
+    border: 1px solid #334155;
+    margin-bottom: 0.5rem;
+}
+
+.filtro-input {
+    background: #0F172A;
+    border: 1px solid #334155;
+    border-radius: 40px;
+    padding: 0.6rem 1rem;
+    color: #E2E8F0;
+    font-family: 'Inter', sans-serif;
+    font-size: 0.9rem;
+    transition: all 0.2s;
+}
+
+.filtro-input:focus {
+    outline: none;
+    border-color: #4CAF50;
+    box-shadow: 0 0 0 2px rgba(76, 175, 80, 0.2);
+}
+
+.btn-sm {
+    padding: 0.5rem 1.2rem;
+    font-size: 0.85rem;
+}
+
+/* Formulario de bitácora */
+.form-card {
+    background: #1E293B;
+    border-radius: 32px;
+    padding: 1.8rem;
+    margin-bottom: 1.5rem;
+    border: 1px solid #334155;
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.3);
+}
+
+.form-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+    gap: 1.5rem;
+}
+
+.form-group {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.form-group label {
+    font-weight: 600;
+    color: #CBD5E1;
+    font-size: 0.85rem;
+}
+
+.form-group input,
+.form-group textarea,
+.form-group select {
+    background: #0F172A;
+    border: 1px solid #334155;
+    border-radius: 20px;
+    padding: 0.8rem 1rem;
+    color: #E2E8F0;
+    font-family: 'Inter', sans-serif;
+    transition: all 0.2s;
+}
+
+.form-group input:focus,
+.form-group textarea:focus,
+.form-group select:focus {
+    outline: none;
+    border-color: #4CAF50;
+    box-shadow: 0 0 0 3px rgba(76, 175, 80, 0.2);
+}
+
+/* Tarjetas de bitácoras */
+.bitacoras-layout {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(360px, 1fr));
+    gap: 1.5rem;
+}
+
+.data-card {
+    background: #1E293B;
+    border-radius: 28px;
+    padding: 1.2rem;
+    border: 1px solid #334155;
+    transition: transform 0.2s, box-shadow 0.2s;
+}
+
+.data-card:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 12px 28px rgba(0, 191, 165, 0.15);
+    border-color: #4CAF50;
+}
+
+.card-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+    margin-bottom: 0.75rem;
+}
+
+.card-title h4 {
+    font-size: 1.1rem;
+    font-weight: 700;
+    color: #E2E8F0;
+    margin: 0;
+}
+
+.card-caption {
+    font-size: 0.8rem;
+    color: #94A3B8;
+    margin: 0.2rem 0 0;
+}
+
+.badge-pill {
+    padding: 0.3rem 0.8rem;
+    border-radius: 40px;
+    font-size: 0.7rem;
+    font-weight: 600;
+    text-transform: uppercase;
+    background: rgba(76, 175, 80, 0.2);
+    color: #A5F3C3;
+}
+
+.card-meta {
+    font-size: 0.75rem;
+    color: #64748B;
+    margin-bottom: 1rem;
+}
+
+.details-list {
+    display: flex;
+    flex-direction: column;
+    gap: 0.75rem;
+}
+
+.detail-item dt {
+    font-weight: 700;
+    color: #CBD5E1;
+    font-size: 0.85rem;
+}
+
+.detail-item dd {
+    margin: 0;
+    color: #E2E8F0;
+    font-size: 0.9rem;
+    line-height: 1.4;
+}
+
+.metrics-row {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.5rem;
+    margin: 0.75rem 0;
+}
+
+.metric-chip {
+    background: #0F172A;
+    border: 1px solid #334155;
+    border-radius: 40px;
+    padding: 0.3rem 0.8rem;
+    font-size: 0.75rem;
+    color: #A5F3C3;
+}
+
+.card-actions {
+    display: flex;
+    justify-content: flex-end;
+    margin-top: 1rem;
+}
+
+/* Paginación */
+.pagination {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 1rem;
+    margin-top: 2rem;
+}
+
+.page-btn {
+    background: #1E293B;
+    border: 1px solid #334155;
+    border-radius: 40px;
+    padding: 0.4rem 1rem;
+    font-size: 1rem;
+    cursor: pointer;
+    color: #E2E8F0;
+    transition: all 0.2s;
+}
+
+.page-btn:hover:not(:disabled) {
+    background: #4CAF50;
+    color: white;
+    border-color: #4CAF50;
+}
+
+.page-btn:disabled {
+    opacity: 0.4;
+    cursor: not-allowed;
+}
+
+.page-info {
+    font-size: 0.9rem;
+    color: #94A3B8;
+}
+
+/* Alertas */
+.alert {
+    background: rgba(76, 175, 80, 0.1);
+    border-left: 4px solid #4CAF50;
+    padding: 1rem;
+    border-radius: 16px;
+    color: #A5F3C3;
+    margin: 0.5rem 0;
+}
+
+.alert.error {
+    background: rgba(239, 68, 68, 0.1);
+    border-left-color: #EF4444;
+    color: #F87171;
+}
+
+.placeholder-card {
+    background: #1E293B;
+    border-radius: 28px;
+    padding: 2rem;
+    text-align: center;
+    border: 1px solid #334155;
+    color: #94A3B8;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+    .filtros-bar {
+        flex-direction: column;
+        align-items: stretch;
+        border-radius: 24px;
+    }
+
+    .bitacoras-layout {
+        grid-template-columns: 1fr;
+    }
+
+    .section-header {
+        flex-direction: column;
+        align-items: flex-start;
+    }
+}
