+++ frontend/src/app/doctor-dashboard/components/doctor-bitacoras/doctor-bitacoras.component.html
@@ -1,0 +1,146 @@
+<div class="bitacoras-wrapper">
+
+<div class="section-header">
+  <h2>Bitácoras Médicas</h2>
+  <div class="section-header-actions">
+    <button class="btn-csv" (click)="exportBitacorasCSV()" [disabled]="bitacoras.length === 0" title="Exportar a CSV">
+      ⬇ CSV
+    </button>
+    <button class="btn-primary" (click)="toggleBitacoraForm()">
+      {{ showBitacoraForm ? 'Cerrar formulario' : 'Registrar Bitácora' }}
+    </button>
+  </div>
+</div>
+
+<div *ngIf="bitacoraMessage" class="alert">{{ bitacoraMessage }}</div>
+
+<div class="filtros-bar">
+  <input type="date" [(ngModel)]="filtrosBitacora.fecha_desde" placeholder="Desde" class="filtro-input" />
+  <input type="date" [(ngModel)]="filtrosBitacora.fecha_hasta" placeholder="Hasta" class="filtro-input" />
+  <input type="text" [(ngModel)]="filtrosBitacora.alumno" placeholder="Buscar alumno..." class="filtro-input" />
+  <button class="btn-primary btn-sm" (click)="aplicarFiltrosBitacora()">Filtrar</button>
+  <button class="btn-secondary btn-sm" (click)="limpiarFiltrosBitacora()">Limpiar</button>
+</div>
+
+<div *ngIf="showBitacoraForm" class="form-card">
+  <form #bitacoraForm="ngForm" (ngSubmit)="onCreateBitacora(bitacoraForm)">
+    <div class="form-grid">
+      <div class="form-group">
+        <label for="citaAsociada">Cita atendida</label>
+        <select id="citaAsociada" name="cita_id" [(ngModel)]="bitacoraFormData.cita_id" required
+          [disabled]="editingBitacoraId !== null">
+          <option value="" disabled selected>Selecciona una cita atendida</option>
+          <option *ngFor="let cita of availableCitasForBitacora" [value]="cita.id">
+            {{ formatDateDisplay(cita.fecha_cita) }} - {{ cita.hora_cita }} hrs ({{ cita.alumno?.nombre }} {{ cita.alumno?.apellido }})
+          </option>
+        </select>
+      </div>
+      <div class="form-group">
+        <label for="diagnostico">Diagnóstico</label>
+        <textarea id="diagnostico" name="diagnostico" [(ngModel)]="bitacoraFormData.diagnostico" rows="2" required
+          placeholder="Describe el diagnóstico"></textarea>
+      </div>
+      <div class="form-group">
+        <label for="tratamiento">Tratamiento</label>
+        <textarea id="tratamiento" name="tratamiento" [(ngModel)]="bitacoraFormData.tratamiento" rows="2" required
+          placeholder="Describe el tratamiento"></textarea>
+      </div>
+    </div>
+
+    <div class="form-grid small-grid">
+      <div class="form-group">
+        <label for="peso">Peso</label>
+        <input type="text" id="peso" name="peso" [(ngModel)]="bitacoraFormData.peso" required placeholder="Ej. 70 kg" />
+      </div>
+      <div class="form-group">
+        <label for="altura">Altura</label>
+        <input type="text" id="altura" name="altura" [(ngModel)]="bitacoraFormData.altura" required placeholder="Ej. 1.70 m" />
+      </div>
+      <div class="form-group">
+        <label for="temperatura">Temperatura</label>
+        <input type="text" id="temperatura" name="temperatura" [(ngModel)]="bitacoraFormData.temperatura" required placeholder="Ej. 36.5 °C" />
+      </div>
+      <div class="form-group">
+        <label for="presion">Presión arterial</label>
+        <input type="text" id="presion" name="presion_arterial" [(ngModel)]="bitacoraFormData.presion_arterial" required placeholder="Ej. 120/80" />
+      </div>
+    </div>
+
+    <div class="form-group">
+      <label for="observaciones">Observaciones adicionales</label>
+      <textarea id="observaciones" name="observaciones" [(ngModel)]="bitacoraFormData.observaciones" rows="3" required
+        placeholder="Notas relevantes de la consulta"></textarea>
+    </div>
+
+    <button class="btn-primary" type="submit" [disabled]="isSubmittingBitacora">
+      {{ isSubmittingBitacora ? 'Guardando...' : 'Guardar bitácora' }}
+    </button>
+  </form>
+</div>
+
+<div class="bitacoras-container">
+  <ng-container *ngIf="!isLoadingBitacoras; else loadingBitacoras">
+    <div *ngIf="bitacorasError" class="alert error">{{ bitacorasError }}</div>
+
+    <div *ngIf="!bitacorasError && bitacoras.length === 0" class="placeholder-card">
+      <h3>No hay bitácoras registradas aún</h3>
+      <p>Registra la información de la consulta para que el alumno pueda consultarla.</p>
+    </div>
+
+    <div *ngIf="bitacoras.length > 0" class="bitacoras-layout">
+      <article *ngFor="let bitacora of pagedBitacoras" class="data-card bitacora-card">
+        <header class="card-header">
+          <div class="card-title">
+            <h4>{{ bitacora.alumno?.nombre }} {{ bitacora.alumno?.apellido }}</h4>
+            <p class="card-caption" *ngIf="bitacora.cita">
+              {{ formatDateDisplay(bitacora.cita.fecha_cita) }} · {{ formatTime(bitacora.cita.hora_cita) }} hrs
+            </p>
+          </div>
+          <span class="badge-pill badge-success">Consulta atendida</span>
+        </header>
+
+        <div class="card-meta">
+          <span class="meta-item">Registrada el {{ formatBitacoraDate(bitacora.created_at) }}</span>
+        </div>
+
+        <div class="card-body">
+          <dl class="details-list">
+            <div class="detail-item" *ngIf="bitacora.diagnostico">
+              <dt>Diagnóstico</dt>
+              <dd>{{ bitacora.diagnostico }}</dd>
+            </div>
+            <div class="detail-item" *ngIf="bitacora.tratamiento">
+              <dt>Tratamiento</dt>
+              <dd>{{ bitacora.tratamiento }}</dd>
+            </div>
+            <div class="detail-item" *ngIf="bitacora.observaciones">
+              <dt>Observaciones</dt>
+              <dd>{{ bitacora.observaciones }}</dd>
+            </div>
+          </dl>
+
+          <div class="metrics-row"
+            *ngIf="bitacora.peso || bitacora.altura || bitacora.temperatura || bitacora.presion_arterial">
+            <span class="metric-chip" *ngIf="bitacora.peso">⚖️ Peso: {{ bitacora.peso }}</span>
+            <span class="metric-chip" *ngIf="bitacora.altura">📏 Altura: {{ bitacora.altura }}</span>
+            <span class="metric-chip" *ngIf="bitacora.temperatura">🌡️ Temp: {{ bitacora.temperatura }}</span>
+            <span class="metric-chip" *ngIf="bitacora.presion_arterial">🩺 Presión: {{ bitacora.presion_arterial }}</span>
+          </div>
+
+          <div class="card-actions">
+            <button class="btn-secondary" type="button" (click)="startEditBitacora(bitacora)">Editar bitácora</button>
+          </div>
+        </div>
+      </article>
+
+      <div class="pagination" *ngIf="totalPagesBitacoras > 1">
+        <button class="page-btn" (click)="prevPageBitacoras()" [disabled]="currentPageBitacoras === 1">‹</button>
+        <span class="page-info">{{ currentPageBitacoras }} / {{ totalPagesBitacoras }}</span>
+        <button class="page-btn" (click)="nextPageBitacoras()" [disabled]="currentPageBitacoras === totalPagesBitacoras">›</button>
+      </div>
+    </div>
+  </ng-container>
+  <ng-template #loadingBitacoras>
+    <div class="placeholder-card">Cargando bitácoras...</div>
+  </ng-template>
+</div>
