+++ frontend/src/app/doctor-dashboard/components/doctor-header/doctor-header.component.ts
@@ -1,0 +1,16 @@
+import { Component, Input, Output, EventEmitter } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@Component({
+  selector: 'app-doctor-header',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './doctor-header.component.html'
+})
+export class DoctorHeaderComponent {
+  @Input() doctorName = 'Doctor';
+  @Input() activeSection = 'inicio';
+  @Input() totalPendientes = 0;
+  @Output() sectionChange = new EventEmitter<string>();
+  @Output() logoutEvent = new EventEmitter<void>();
+}
