+++ frontend/src/app/doctor-dashboard/components/doctor-inicio/doctor-inicio.component.css
@@ -1,0 +1,206 @@
+/* ===== DOCTOR INICIO - ESTILOS MEJORADOS ===== */
+.inicio-container {
+    display: flex;
+    flex-direction: column;
+    gap: 2rem;
+    max-width: 1200px;
+    margin: 0 auto;
+}
+
+/* Encabezado de bienvenida */
+.welcome-header {
+    text-align: center;
+    margin-bottom: 0.5rem;
+}
+
+.welcome-header h2 {
+    font-size: 2rem;
+    font-weight: 700;
+    background: linear-gradient(135deg, #E2E8F0, #A5F3C3);
+    background-clip: text;
+    -webkit-background-clip: text;
+    color: transparent;
+    margin-bottom: 0.5rem;
+}
+
+.welcome-header p {
+    font-size: 1rem;
+    color: #94A3B8;
+}
+
+/* Cuadrícula de estadísticas (más espaciada) */
+.stats-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(280px, 1fr));
+    gap: 1.8rem;
+    margin-bottom: 1rem;
+}
+
+.stat-card {
+    background: #1E293B;
+    border-radius: 28px;
+    padding: 1.5rem;
+    display: flex;
+    align-items: center;
+    gap: 1.2rem;
+    border: 1px solid #334155;
+    transition: transform 0.2s, box-shadow 0.2s;
+    box-shadow: 0 8px 20px rgba(0, 0, 0, 0.3);
+}
+
+.stat-card:hover {
+    transform: translateY(-4px);
+    box-shadow: 0 16px 30px rgba(0, 191, 165, 0.2);
+    border-color: #4CAF50;
+}
+
+.stat-icon {
+    font-size: 2.8rem;
+    background: #0F172A;
+    width: 70px;
+    height: 70px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 50%;
+    border: 1px solid #334155;
+}
+
+.stat-content {
+    flex: 1;
+}
+
+.stat-content h3 {
+    font-size: 1rem;
+    font-weight: 500;
+    color: #94A3B8;
+    margin-bottom: 0.5rem;
+    letter-spacing: 0.5px;
+}
+
+.stat-value {
+    font-size: 2.5rem;
+    font-weight: 800;
+    color: #66BB6A;
+    margin: 0.2rem 0;
+    line-height: 1;
+}
+
+.stat-detail {
+    font-size: 0.75rem;
+    color: #64748B;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+}
+
+/* Próxima cita (destacada) */
+.proxima-cita-wrapper {
+    margin-top: 1rem;
+}
+
+.proxima-cita-card {
+    background: linear-gradient(135deg, #1E293B 0%, #0F172A 100%);
+    border-radius: 32px;
+    padding: 1.8rem;
+    border: 1px solid #334155;
+    box-shadow: 0 12px 28px rgba(0, 0, 0, 0.4);
+    transition: all 0.2s;
+}
+
+.proxima-cita-card:not(.empty) {
+    border-left: 6px solid #4CAF50;
+    background: linear-gradient(135deg, #1E293B, #0F172A);
+}
+
+.proxima-cita-card.empty {
+    text-align: center;
+    background: #1E293B;
+    opacity: 0.8;
+}
+
+.proxima-header {
+    display: flex;
+    align-items: center;
+    gap: 0.8rem;
+    margin-bottom: 1.5rem;
+    border-bottom: 1px solid #334155;
+    padding-bottom: 0.8rem;
+}
+
+.proxima-icon {
+    font-size: 1.8rem;
+}
+
+.proxima-header h3 {
+    font-size: 1.4rem;
+    font-weight: 700;
+    color: #E2E8F0;
+    margin: 0;
+}
+
+.proxima-cita-info {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+    gap: 1.2rem;
+}
+
+.proxima-dato {
+    display: flex;
+    flex-direction: column;
+    gap: 0.3rem;
+}
+
+.proxima-label {
+    font-size: 0.75rem;
+    font-weight: 600;
+    text-transform: uppercase;
+    letter-spacing: 0.5px;
+    color: #94A3B8;
+}
+
+.proxima-valor {
+    font-size: 1rem;
+    font-weight: 500;
+    color: #CBD5E1;
+    word-break: break-word;
+}
+
+.proxima-valor.codigo {
+    font-family: monospace;
+    background: #0F172A;
+    padding: 0.2rem 0.5rem;
+    border-radius: 12px;
+    display: inline-block;
+    font-size: 0.9rem;
+    letter-spacing: 1px;
+    border: 1px solid #334155;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+    .stats-grid {
+        gap: 1rem;
+    }
+
+    .stat-card {
+        padding: 1rem;
+    }
+
+    .stat-icon {
+        width: 55px;
+        height: 55px;
+        font-size: 2rem;
+    }
+
+    .stat-value {
+        font-size: 1.8rem;
+    }
+
+    .proxima-cita-info {
+        grid-template-columns: 1fr;
+    }
+
+    .proxima-header h3 {
+        font-size: 1.2rem;
+    }
+}
