+++ frontend/src/app/doctor-dashboard/doctor-dashboard.component.ts
@@ -1,1440 +1,52 @@
-import { Component, OnDestroy, OnInit, AfterViewChecked, ElementRef, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewEncapsulation } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RouterModule, Router } from '@angular/router';
-import { FormsModule, NgForm } from '@angular/forms';
-import { Subject, of } from 'rxjs';
-import { catchError, finalize, takeUntil } from 'rxjs/operators';
-import { Chart, registerables } from 'chart.js';
+import { Router } from '@angular/router';
 import { AuthService } from '../services/auth.service';
-import { Cita, CitaService, CreateCitaPayload, AvailabilityStatus, CitaAvailabilityDay } from '../services/cita.service';
-import { Bitacora, BitacoraService, CreateBitacoraPayload } from '../services/bitacora.service';
-import { Receta, RecetaService, CreateRecetaPayload } from '../services/receta.service';
-import { PreEvaluacionIAService, PreEvaluacion } from '../services/pre-evaluacion-ia.service';
-import { IaPriorityService, ClasificacionPrioridad, ResumenPrioridad } from '../services/ia-priority.service';
-import { EstadisticasService, Estadisticas } from '../services/estadisticas.service';
-import { ConsultaService } from '../services/consulta.service';
-import { PerfilMedicoService, PerfilMedico, ConsultaHistorial } from '../services/perfil-medico.service';
-
-Chart.register(...registerables);
-
-type CalendarAvailability = 'none' | AvailabilityStatus;
-
-interface DayAvailabilityRecord {
-  takenSlots: Set<string>;
-  status: AvailabilityStatus;
-  color?: string;
-  label?: string | null;
-}
-
-interface CalendarDay {
-  date: string;
-  label: number;
-  isCurrentMonth: boolean;
-  isToday: boolean;
-  isPast: boolean;
-  availability: CalendarAvailability;
-  color?: string | null;
-  labelText?: string | null;
-}
+import { DoctorHeaderComponent } from './components/doctor-header/doctor-header.component';
+import { DoctorInicioComponent } from './components/doctor-inicio/doctor-inicio.component';
+import { DoctorCitasComponent } from './components/doctor-citas/doctor-citas.component';
+import { DoctorBitacorasComponent } from './components/doctor-bitacoras/doctor-bitacoras.component';
+import { DoctorRecetasComponent } from './components/doctor-recetas/doctor-recetas.component';
+import { DoctorPreEvaluacionesComponent } from './components/doctor-pre-evaluaciones/doctor-pre-evaluaciones.component';
+import { DoctorIaPrioridadComponent } from './components/doctor-ia-prioridad/doctor-ia-prioridad.component';
+import { DoctorEstadisticasComponent } from './components/doctor-estadisticas/doctor-estadisticas.component';
 
 @Component({
   selector: 'app-doctor-dashboard',
   standalone: true,
-  imports: [CommonModule, RouterModule, FormsModule],
+  encapsulation: ViewEncapsulation.None,
+  imports: [
+    CommonModule,
+    DoctorHeaderComponent,
+    DoctorInicioComponent,
+    DoctorCitasComponent,
+    DoctorBitacorasComponent,
+    DoctorRecetasComponent,
+    DoctorPreEvaluacionesComponent,
+    DoctorIaPrioridadComponent,
+    DoctorEstadisticasComponent
+  ],
   templateUrl: './doctor-dashboard.component.html',
   styleUrls: ['./doctor-dashboard.component.css']
 })
-
-export class DoctorDashboardComponent implements OnInit, OnDestroy, AfterViewChecked {
-  @ViewChild('barCanvas') barCanvas!: ElementRef<HTMLCanvasElement>;
-  @ViewChild('doughnutCanvas') doughnutCanvas!: ElementRef<HTMLCanvasElement>;
-  activeSection: string = 'inicio';
-  doctorName: string = 'Doctor';
-  today: string = this.formatDate(new Date());
-
-  citas: Cita[] = [];
-  isLoadingCitas = false;
-  citasError: string | null = null;
-
-  pendingCitas: Cita[] = [];
-  handledCitas: Cita[] = [];
-  searchCitas = '';
-
-  showCreateForm = false;
-  isSubmitting = false;
-  submitMessage: string | null = null;
-
-  createFormData: Partial<CreateCitaPayload & { numero_control: string }> = {
-    fecha_cita: '',
-    hora_cita: '',
-    motivo: '',
-    numero_control: ''
-  };
-  readonly timeSlots: string[] = this.generateTimeSlots();
-  readonly totalSlotsPerDay = this.timeSlots.length;
-  readonly weekDays: string[] = ['Dom', 'Lun', 'Mar', 'Mié', 'Jue', 'Vie', 'Sáb'];
-  currentMonth: Date = this.startOfMonth(new Date());
-  calendarWeeks: CalendarDay[][] = [];
-  availabilityMap: Map<string, DayAvailabilityRecord> = new Map();
-  isLoadingAvailability = false;
-
-  todayAppointments = 0;
-  patientsAttended = 0;
-  pendingBitacoras = 0;
-  proximaCita: Cita | null = null;
-
-  // IA Prioridad
-  prioridadResumen: ResumenPrioridad | null = null;
-  isLoadingPrioridad = false;
-  prioridadError: string | null = null;
-
-  // Perfil médico del alumno (modal)
-  showPerfilAlumnoModal = false;
-  perfilAlumnoData: PerfilMedico | null = null;
-  historialAlumno: ConsultaHistorial[] = [];
-  isLoadingPerfilAlumno = false;
-  historialAlumnoExpandido: number | null = null;
-
-  // Formulario de consulta
-  showConsultaForm = false;
-  consultaCitaId: number | null = null;
-  consultaCitaLabel = '';
-  consultaForm = { diagnostico: '', tratamiento: '', observaciones: '' };
-  consultaMsg: string | null = null;
-  isSubmittingConsulta = false;
-
-  // Paginación
-  readonly PAGE_SIZE = 10;
-  currentPageBitacoras = 1;
-  currentPageRecetas = 1;
-
-  get pagedBitacoras(): Bitacora[] {
-    const start = (this.currentPageBitacoras - 1) * this.PAGE_SIZE;
-    return this.bitacoras.slice(start, start + this.PAGE_SIZE);
-  }
-  get totalPagesBitacoras(): number { return Math.ceil(this.bitacoras.length / this.PAGE_SIZE) || 1; }
-
-  get pagedRecetas(): Receta[] {
-    const start = (this.currentPageRecetas - 1) * this.PAGE_SIZE;
-    return this.recetas.slice(start, start + this.PAGE_SIZE);
-  }
-  get totalPagesRecetas(): number { return Math.ceil(this.recetas.length / this.PAGE_SIZE) || 1; }
-
-  // Estadísticas
-  estadisticas: Estadisticas | null = null;
-  isLoadingEstadisticas = false;
-  estadisticasError: string | null = null;
-  private barChart: Chart | null = null;
-  private doughnutChart: Chart | null = null;
-  private chartsRendered = false;
-
-  bitacoras: Bitacora[] = [];
-  isLoadingBitacoras = false;
-  bitacorasError: string | null = null;
-  filtrosBitacora: { fecha_desde: string; fecha_hasta: string; alumno: string } = { fecha_desde: '', fecha_hasta: '', alumno: '' };
-  showBitacoraForm = false;
-  isSubmittingBitacora = false;
-  bitacoraMessage: string | null = null;
-  editingBitacoraId: number | null = null;
-  bitacoraFormData: Partial<CreateBitacoraPayload> = {
-    cita_id: undefined,
-    diagnostico: '',
-    tratamiento: '',
-    observaciones: '',
-    peso: '',
-    altura: '',
-    temperatura: '',
-    presion_arterial: ''
-  };
-
-  recetas: Receta[] = [];
-  isLoadingRecetas = false;
-  recetasError: string | null = null;
-  showRecetaForm = false;
-  isSubmittingReceta = false;
-  recetaMessage: string | null = null;
-  editingRecetaId: number | null = null;
-  recetaFormData: Partial<CreateRecetaPayload> = {
-    cita_id: undefined,
-    medicamentos: '',
-    indicaciones: '',
-    fecha_emision: this.formatDate(new Date())
-  };
-
-  preEvaluaciones: PreEvaluacion[] = [];
-  preEvaluacionesHistorial: PreEvaluacion[] = [];
-  isLoadingPreEvaluaciones = false;
-  preEvaluacionesError: string | null = null;
-  showPreEvaluacionModal = false;
-  selectedPreEvaluacion: PreEvaluacion | null = null;
-  isSubmittingValidacion = false;
-  comentarioValidacion: string = '';
+export class DoctorDashboardComponent implements OnInit {
+  activeSection = 'inicio';
+  doctorName = 'Doctor';
   totalPendientes = 0;
 
-  private destroy$ = new Subject<void>();
-
-  constructor(
-    private router: Router,
-    private authService: AuthService,
-    private citaService: CitaService,
-    private bitacoraService: BitacoraService,
-    private recetaService: RecetaService,
-    private preEvaluacionIAService: PreEvaluacionIAService,
-    private iaPriorityService: IaPriorityService,
-    private estadisticasService: EstadisticasService,
-    private consultaService: ConsultaService,
-    private perfilMedicoService: PerfilMedicoService
-  ) {}
+  constructor(private router: Router, private authService: AuthService) {}
 
   ngOnInit(): void {
     const user = this.authService.getCurrentUser();
     this.doctorName = user ? `${user.nombre} ${user.apellido}` : 'Doctor';
-    this.buildCalendar();
-    this.loadAvailability();
-    this.loadCitas();
-    this.loadBitacoras();
-    this.loadRecetas();
   }
 
-  ngOnDestroy(): void {
-    this.destroy$.next();
-    this.destroy$.complete();
-    this.barChart?.destroy();
-    this.doughnutChart?.destroy();
-  }
-
-  ngAfterViewChecked(): void {
-    if (this.activeSection === 'estadisticas' && this.estadisticas && !this.chartsRendered) {
-      if (this.barCanvas && this.doughnutCanvas) {
-        this.renderCharts();
-        this.chartsRendered = true;
-      }
-    }
-  }
-
-  get filteredPendingCitas(): Cita[] {
-    return this.filterCitas(this.pendingCitas);
-  }
-
-  get filteredHandledCitas(): Cita[] {
-    return this.filterCitas(this.handledCitas);
-  }
-
-  private filterCitas(citas: Cita[]): Cita[] {
-    const q = this.searchCitas.trim().toLowerCase();
-    if (!q) return citas;
-    return citas.filter(c => {
-      const nombre = `${c.alumno?.nombre ?? ''} ${c.alumno?.apellido ?? ''}`.toLowerCase();
-      const nc = (c.alumno?.numero_control ?? '').toLowerCase();
-      const motivo = (c.motivo ?? '').toLowerCase();
-      return nombre.includes(q) || nc.includes(q) || motivo.includes(q);
-    });
-  }
-
-  setActiveSection(section: string) {
+  setActiveSection(section: string): void {
     this.activeSection = section;
-    if (section === 'citas') {
-      this.loadCitas();
-      this.loadAvailability();
-    }
-    if (section === 'bitacoras') {
-      this.loadBitacoras();
-    }
-    if (section === 'recetas') {
-      this.loadRecetas();
-    }
-    if (section === 'pre-evaluaciones') {
-      this.loadPreEvaluaciones();
-    }
-    if (section === 'ia-prioridad') {
-      this.loadPrioridad();
-    }
-    if (section === 'estadisticas') {
-      this.chartsRendered = false;
-      this.loadEstadisticas();
-    }
   }
 
-  toggleCreateForm(): void {
-    this.showCreateForm = !this.showCreateForm;
-    this.submitMessage = null;
-    if (this.showCreateForm) {
-      this.loadAvailability();
-    } else {
-      this.resetForm();
-    }
-  }
-
-  toggleBitacoraForm(): void {
-    this.showBitacoraForm = !this.showBitacoraForm;
-    this.bitacoraMessage = null;
-    if (!this.showBitacoraForm) {
-      this.editingBitacoraId = null;
-      this.resetBitacoraForm();
-    }
-  }
-
-  toggleRecetaForm(): void {
-    this.showRecetaForm = !this.showRecetaForm;
-    this.recetaMessage = null;
-    if (!this.showRecetaForm) {
-      this.editingRecetaId = null;
-      this.resetRecetaForm();
-    }
-  }
-
-  get calendarLabel(): string {
-    return new Intl.DateTimeFormat('es-MX', {
-      month: 'long',
-      year: 'numeric'
-    }).format(this.currentMonth);
-  }
-
-  get isCurrentMonth(): boolean {
-    const now = new Date();
-    return this.currentMonth.getFullYear() === now.getFullYear() &&
-           this.currentMonth.getMonth() === now.getMonth();
-  }
-
-  get hasAvailableSlotsForSelectedDate(): boolean {
-    if (!this.createFormData.fecha_cita) {
-      return false;
-    }
-
-    const record = this.getDayRecord(this.createFormData.fecha_cita);
-    if (record?.status === 'full') {
-      return false;
-    }
-
-    const takenSlots = record?.takenSlots ?? new Set<string>();
-    return this.timeSlots.some(slot => !takenSlots.has(this.normalizeTime(slot)));
-  }
-
-  changeMonth(direction: number): void {
-    this.currentMonth = this.startOfMonth(
-      new Date(this.currentMonth.getFullYear(), this.currentMonth.getMonth() + direction, 1)
-    );
-    this.availabilityMap = new Map<string, DayAvailabilityRecord>();
-    this.buildCalendar();
-    this.loadAvailability();
-  }
-
-  selectCalendarDay(day: CalendarDay): void {
-    if (day.isPast || !day.isCurrentMonth || day.availability === 'full') {
-      return;
-    }
-
-    this.onDateChange(day.date);
-  }
-
-  selectTimeSlot(slot: string): void {
-    if (!this.createFormData.fecha_cita || this.isSlotUnavailable(slot)) {
-      return;
-    }
-
-    this.createFormData.hora_cita = this.normalizeTime(slot);
-    this.submitMessage = null;
-  }
-
-  
-  onCreateCita(form: NgForm): void {
-    if (form.invalid || !this.createFormData.numero_control?.trim()) {
-      this.submitMessage = 'Ingresa el número de control del alumno.';
-      return;
-    }
-
-    const normalizedTime = this.normalizeTime(this.createFormData.hora_cita!);
-
-    if (this.isSlotUnavailable(normalizedTime)) {
-      this.submitMessage = 'La hora seleccionada ya está ocupada. Elige otro bloque disponible.';
-      return;
-    }
-
-    const payload: CreateCitaPayload = {
-      fecha_cita: this.createFormData.fecha_cita!,
-      hora_cita: normalizedTime,
-      motivo: this.createFormData.motivo || undefined,
-      numero_control: this.createFormData.numero_control.trim()
-    };
-
-    this.isSubmitting = true;
-    this.submitMessage = null;
-
-    this.citaService.createCita(payload)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          const errores = error?.error?.errors;
-          if (errores && typeof errores === 'object') {
-            const firstKey = Object.keys(errores)[0];
-            const mensajes = (errores as any)[firstKey];
-            if (Array.isArray(mensajes) && mensajes.length > 0) {
-              this.submitMessage = mensajes[0];
-              return of(null);
-            }
-          }
-          this.submitMessage = error?.error?.message || error?.message || 'No se pudo agendar la cita.';
-          return of(null);
-        }),
-        finalize(() => {
-          this.isSubmitting = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.cita) {
-          this.submitMessage = 'Cita agendada correctamente.';
-          this.loadCitas();
-          this.loadAvailability();
-          this.resetForm();
-          this.showCreateForm = false;
-        }
-      });
-  }
-
-  onCancelCita(cita: Cita): void {
-    if (this.isSubmitting) {
-      return;
-    }
-
-    this.isSubmitting = true;
-    this.citaService.cancelCita(cita.id)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.submitMessage = error?.error?.message || 'No se pudo cancelar la cita.';
-          return of(null);
-        }),
-        finalize(() => {
-          this.isSubmitting = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.cita) {
-          this.submitMessage = 'Cita cancelada correctamente.';
-          this.loadCitas();
-          this.loadAvailability();
-        }
-      });
-  }
-
-  onMarkAsAttended(cita: Cita): void {
-    // Abrir formulario de consulta antes de marcar como atendida
-    this.consultaCitaId = cita.id;
-    this.consultaCitaLabel = `${cita.alumno?.nombre ?? ''} ${cita.alumno?.apellido ?? ''} — ${cita.fecha_cita}`;
-    this.consultaForm = { diagnostico: '', tratamiento: '', observaciones: '' };
-    this.consultaMsg = null;
-    this.showConsultaForm = true;
-  }
-
-  closeConsultaForm(): void {
-    this.showConsultaForm = false;
-    this.consultaCitaId = null;
-    this.consultaMsg = null;
-  }
-
-  submitConsulta(): void {
-    if (!this.consultaCitaId) return;
-    if (!this.consultaForm.diagnostico.trim() || !this.consultaForm.tratamiento.trim()) {
-      this.consultaMsg = 'Diagnóstico y tratamiento son obligatorios.';
-      return;
-    }
-    this.isSubmittingConsulta = true;
-    this.consultaMsg = null;
-    this.consultaService.guardar(this.consultaCitaId, this.consultaForm)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.consultaMsg = error?.error?.message || 'Error al guardar la consulta.';
-          return of(null);
-        }),
-        finalize(() => { this.isSubmittingConsulta = false; })
-      )
-      .subscribe(response => {
-        if (response) {
-          this.consultaMsg = 'Consulta guardada y cita marcada como atendida.';
-          this.loadCitas();
-          this.loadBitacoras();
-          setTimeout(() => this.closeConsultaForm(), 1200);
-        }
-      });
-  }
-
-  onMarkAsNoShow(cita: Cita): void {
-    if (this.isSubmitting) {
-      return;
-    }
-
-    this.isSubmitting = true;
-    this.citaService.markAsNoShow(cita.id)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.submitMessage = error?.error?.message || 'No se pudo marcar como no asistida.';
-          return of(null);
-        }),
-        finalize(() => {
-          this.isSubmitting = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.cita) {
-          this.submitMessage = 'Cita marcada como no asistida.';
-          this.loadCitas();
-        }
-      });
-  }
-
-  onCreateBitacora(form: NgForm): void {
-    if (form.invalid) {
-      if (!this.bitacoraFormData.cita_id) {
-        this.bitacoraMessage = 'Selecciona la cita atendida correspondiente.';
-      } else {
-        this.bitacoraMessage = 'Por favor completa todos los campos obligatorios de la bitácora.';
-      }
-      return;
-    }
-
-    const payload: CreateBitacoraPayload = {
-      cita_id: Number(this.bitacoraFormData.cita_id),
-      diagnostico: this.bitacoraFormData.diagnostico || undefined,
-      tratamiento: this.bitacoraFormData.tratamiento || undefined,
-      observaciones: this.bitacoraFormData.observaciones || undefined,
-      peso: this.bitacoraFormData.peso || undefined,
-      altura: this.bitacoraFormData.altura || undefined,
-      temperatura: this.bitacoraFormData.temperatura || undefined,
-      presion_arterial: this.bitacoraFormData.presion_arterial || undefined
-    };
-
-    this.isSubmittingBitacora = true;
-    this.bitacoraMessage = null;
-
-    const request$ = this.editingBitacoraId
-      ? this.bitacoraService.updateBitacora(this.editingBitacoraId, payload)
-      : this.bitacoraService.createBitacora(payload);
-
-    request$
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          const errores = error?.error?.errors;
-          if (errores && typeof errores === 'object') {
-            const firstKey = Object.keys(errores)[0];
-            const mensajes = (errores as any)[firstKey];
-            if (Array.isArray(mensajes) && mensajes.length > 0) {
-              this.bitacoraMessage = mensajes[0];
-              return of(null);
-            }
-          }
-
-          this.bitacoraMessage = error?.error?.message || 'No se pudo registrar la bitácora.';
-          return of(null);
-        }),
-        finalize(() => {
-          this.isSubmittingBitacora = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.bitacora) {
-          this.bitacoraMessage = this.editingBitacoraId
-            ? 'Bitácora actualizada correctamente.'
-            : 'Bitácora registrada correctamente.';
-          this.loadBitacoras();
-          this.editingBitacoraId = null;
-          this.resetBitacoraForm();
-          this.showBitacoraForm = false;
-        }
-      });
-  }
-
-  onCreateReceta(form: NgForm): void {
-    if (form.invalid) {
-      if (!this.recetaFormData.cita_id) {
-        this.recetaMessage = 'Selecciona la cita atendida correspondiente.';
-      } else if (!this.recetaFormData.medicamentos?.trim()) {
-        this.recetaMessage = 'Captura los medicamentos recetados.';
-      } else if (!this.recetaFormData.indicaciones?.trim()) {
-        this.recetaMessage = 'Captura las indicaciones para el paciente.';
-      } else if (!this.recetaFormData.fecha_emision) {
-        this.recetaMessage = 'Selecciona la fecha de emisión de la receta.';
-      } else {
-        this.recetaMessage = 'Por favor completa todos los campos obligatorios de la receta.';
-      }
-      return;
-    }
-
-    const payload: CreateRecetaPayload = {
-      cita_id: Number(this.recetaFormData.cita_id),
-      medicamentos: this.recetaFormData.medicamentos!.trim(),
-      indicaciones: this.recetaFormData.indicaciones?.trim() || undefined,
-      fecha_emision: this.recetaFormData.fecha_emision || this.formatDate(new Date())
-    };
-
-    this.isSubmittingReceta = true;
-    this.recetaMessage = null;
-
-    const request$ = this.editingRecetaId
-      ? this.recetaService.updateReceta(this.editingRecetaId, payload)
-      : this.recetaService.createReceta(payload);
-
-    request$
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          const errores = error?.error?.errors;
-          if (errores && typeof errores === 'object') {
-            const firstKey = Object.keys(errores)[0];
-            const mensajes = (errores as any)[firstKey];
-            if (Array.isArray(mensajes) && mensajes.length > 0) {
-              this.recetaMessage = mensajes[0];
-              return of(null);
-            }
-          }
-
-          this.recetaMessage = error?.error?.message || 'No se pudo registrar la receta.';
-          return of(null);
-        }),
-        finalize(() => {
-          this.isSubmittingReceta = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.receta) {
-          this.recetaMessage = this.editingRecetaId
-            ? 'Receta actualizada correctamente.'
-            : 'Receta registrada correctamente.';
-          this.loadRecetas();
-          this.editingRecetaId = null;
-          this.resetRecetaForm();
-          this.showRecetaForm = false;
-        }
-      });
-  }
-
-  logout() {
+  logout(): void {
     this.authService.logout();
     this.router.navigate(['/login']);
-  }
-
-  private loadCitas(): void {
-    this.isLoadingCitas = true;
-    this.citasError = null;
-
-    this.citaService.getCitas()
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.citasError = error?.error?.message || 'No se pudieron obtener las citas.';
-          return of([] as Cita[]);
-        }),
-        finalize(() => {
-          this.isLoadingCitas = false;
-        })
-      )
-      .subscribe(citas => {
-        this.citas = citas;
-        this.updateStats();
-        this.updateCitaGroups();
-      });
-  }
-
-  aplicarFiltrosBitacora(): void {
-    this.currentPageBitacoras = 1;
-    this.loadBitacoras();
-  }
-
-  limpiarFiltrosBitacora(): void {
-    this.filtrosBitacora = { fecha_desde: '', fecha_hasta: '', alumno: '' };
-    this.currentPageBitacoras = 1;
-    this.loadBitacoras();
-  }
-
-  prevPageBitacoras(): void { if (this.currentPageBitacoras > 1) this.currentPageBitacoras--; }
-  nextPageBitacoras(): void { if (this.currentPageBitacoras < this.totalPagesBitacoras) this.currentPageBitacoras++; }
-  prevPageRecetas(): void { if (this.currentPageRecetas > 1) this.currentPageRecetas--; }
-  nextPageRecetas(): void { if (this.currentPageRecetas < this.totalPagesRecetas) this.currentPageRecetas++; }
-
-  exportBitacorasCSV(): void {
-    const headers = ['Fecha Cita', 'Alumno', 'Diagnóstico', 'Tratamiento', 'Observaciones', 'Peso', 'Altura', 'Temperatura', 'Presión Arterial', 'Registrada'];
-    const rows = this.bitacoras.map(b => [
-      b.cita?.fecha_cita ?? '',
-      `${b.alumno?.nombre ?? ''} ${b.alumno?.apellido ?? ''}`.trim(),
-      b.diagnostico ?? '',
-      b.tratamiento ?? '',
-      b.observaciones ?? '',
-      b.peso ?? '',
-      b.altura ?? '',
-      b.temperatura ?? '',
-      b.presion_arterial ?? '',
-      b.created_at ? new Date(b.created_at).toLocaleDateString('es-MX') : ''
-    ].map(v => `"${String(v).replace(/"/g, '""')}"`));
-
-    const csv = [headers.join(','), ...rows.map(r => r.join(','))].join('\n');
-    const blob = new Blob(['\ufeff' + csv], { type: 'text/csv;charset=utf-8;' });
-    const url = URL.createObjectURL(blob);
-    const a = document.createElement('a');
-    a.href = url;
-    a.download = `bitacoras_${new Date().toISOString().slice(0, 10)}.csv`;
-    a.click();
-    URL.revokeObjectURL(url);
-  }
-
-  private loadBitacoras(): void {
-    if (this.isLoadingBitacoras) {
-      return;
-    }
-
-    this.isLoadingBitacoras = true;
-    this.bitacorasError = null;
-
-    const filtros = {
-      fecha_desde: this.filtrosBitacora.fecha_desde || undefined,
-      fecha_hasta: this.filtrosBitacora.fecha_hasta || undefined,
-      alumno: this.filtrosBitacora.alumno || undefined,
-    };
-
-    this.bitacoraService.getBitacoras(filtros)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.bitacorasError = error?.error?.message || 'No se pudieron obtener las bitácoras.';
-          return of([] as Bitacora[]);
-        }),
-        finalize(() => {
-          this.isLoadingBitacoras = false;
-        })
-      )
-      .subscribe(bitacoras => {
-        // Ordenar de la más reciente a la más antigua
-        this.bitacoras = [...bitacoras].sort((a, b) =>
-          new Date(b.created_at).getTime() - new Date(a.created_at).getTime()
-        );
-        this.updateStats();
-      });
-  }
-
-  private loadRecetas(): void {
-    if (this.isLoadingRecetas) {
-      return;
-    }
-
-    this.isLoadingRecetas = true;
-    this.recetasError = null;
-
-    this.recetaService.getRecetas()
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.recetasError = error?.error?.message || 'No se pudieron obtener las recetas.';
-          return of([] as Receta[]);
-        }),
-        finalize(() => {
-          this.isLoadingRecetas = false;
-        })
-      )
-      .subscribe(recetas => {
-        this.recetas = [...recetas].sort((a, b) =>
-          new Date(b.fecha_emision).getTime() - new Date(a.fecha_emision).getTime()
-        );
-        this.currentPageRecetas = 1;
-        this.updateStats();
-      });
-  }
-
-  private resetForm(): void {
-    this.createFormData = {
-      fecha_cita: '',
-      hora_cita: '',
-      motivo: '',
-      numero_control: ''
-    };
-  }
-
-  private resetBitacoraForm(): void {
-    this.bitacoraFormData = {
-      cita_id: undefined,
-      diagnostico: '',
-      tratamiento: '',
-      observaciones: '',
-      peso: '',
-      altura: '',
-      temperatura: '',
-      presion_arterial: ''
-    };
-  }
-
-  public startEditBitacora(bitacora: Bitacora): void {
-    this.showBitacoraForm = true;
-    this.editingBitacoraId = bitacora.id;
-    this.bitacoraMessage = null;
-    this.bitacoraFormData = {
-      cita_id: bitacora.cita_id,
-      diagnostico: bitacora.diagnostico || '',
-      tratamiento: bitacora.tratamiento || '',
-      observaciones: bitacora.observaciones || '',
-      peso: bitacora.peso || '',
-      altura: bitacora.altura || '',
-      temperatura: bitacora.temperatura || '',
-      presion_arterial: bitacora.presion_arterial || ''
-    };
-  }
-
-  private resetRecetaForm(): void {
-    this.recetaFormData = {
-      cita_id: undefined,
-      medicamentos: '',
-      indicaciones: '',
-      fecha_emision: this.formatDate(new Date())
-    };
-  }
-
-  public startEditReceta(receta: Receta): void {
-    this.showRecetaForm = true;
-    this.editingRecetaId = receta.id;
-    this.recetaMessage = null;
-    this.recetaFormData = {
-      cita_id: receta.cita_id,
-      medicamentos: receta.medicamentos,
-      indicaciones: receta.indicaciones || '',
-      fecha_emision: receta.fecha_emision
-    };
-  }
-
-  private updateStats(): void {
-    // Citas del día: solo citas programadas para la fecha actual
-    this.todayAppointments = this.citas.filter(
-      cita => cita.estatus === 'programada' && cita.fecha_cita === this.today
-    ).length;
-
-    // Pacientes atendidos: total de citas marcadas como atendidas
-    this.patientsAttended = this.citas.filter(cita => cita.estatus === 'atendida').length;
-
-    // Citas pendientes: total de citas con estatus "programada",
-    // igual que el conteo que ve el alumno en su panel
-    this.pendingBitacoras = this.citas.filter(cita => cita.estatus === 'programada').length;
-  }
-
-  private updateCitaGroups(): void {
-    const ahora = new Date();
-    this.pendingCitas = this.citas
-      .filter(cita => cita.estatus === 'programada')
-      .sort(
-        (a, b) =>
-          this.toDate(a.fecha_cita, a.hora_cita).getTime() -
-          this.toDate(b.fecha_cita, b.hora_cita).getTime()
-      );
-
-    // Próxima cita: la más cercana en el futuro
-    this.proximaCita = this.pendingCitas.find(
-      cita => this.toDate(cita.fecha_cita, cita.hora_cita) >= ahora
-    ) ?? this.pendingCitas[0] ?? null;
-
-    this.handledCitas = this.citas
-      .filter(cita => cita.estatus !== 'programada')
-      .sort(
-        (a, b) =>
-          this.toDate(b.fecha_cita, b.hora_cita).getTime() -
-          this.toDate(a.fecha_cita, a.hora_cita).getTime()
-      );
-  }
-
-  public formatDateDisplay(fecha: string): string {
-    return this.formatFullDate(fecha);
-  }
-
-  public formatBitacoraDate(fecha: string): string {
-    return this.formatDateTime(new Date(fecha));
-  }
-
-  public formatRecetaDate(fecha: string): string {
-    return new Intl.DateTimeFormat('es-MX', {
-      dateStyle: 'medium'
-    }).format(new Date(fecha));
-  }
-
-  public onRecetaCitaChange(citaId: number | string): void {
-    const id = Number(citaId);
-    const cita = this.citas.find(c => c.id === id);
-    if (cita) {
-      // Usar la fecha de la cita como fecha de emisión por defecto (YYYY-MM-DD)
-      this.recetaFormData.fecha_emision = cita.fecha_cita;
-    }
-  }
-
-  public get availableCitasForBitacora(): Cita[] {
-    const bitacoraCitaIds = new Set(this.bitacoras.map(bitacora => bitacora.cita_id));
-    return this.citas.filter(
-      cita => cita.estatus === 'atendida' && !bitacoraCitaIds.has(cita.id)
-    );
-  }
-
-  public get availableCitasForReceta(): Cita[] {
-    // Si estamos editando una receta, mostramos todas las citas atendidas.
-    // La validación de "una receta por cita" ya está controlada por el backend
-    // y el select está deshabilitado en modo edición.
-    if (this.editingRecetaId !== null) {
-      return this.citas.filter(cita => cita.estatus === 'atendida');
-    }
-
-    // Al crear una receta nueva, solo permitir citas atendidas que aún no tengan receta.
-    const recetaCitaIds = new Set(this.recetas.map(receta => receta.cita_id));
-    return this.citas.filter(
-      cita => cita.estatus === 'atendida' && !recetaCitaIds.has(cita.id)
-    );
-  }
-
-  public formatTime(hora: string): string {
-    const [hours, minutes] = this.normalizeTime(hora).split(':').map(Number);
-    const date = new Date();
-    date.setHours(hours, minutes || 0, 0, 0);
-    return new Intl.DateTimeFormat('es-MX', {
-      hour: '2-digit',
-      minute: '2-digit',
-      hour12: false
-    }).format(date);
-  }
-
-  public isSelectedDate(date: string): boolean {
-    return this.createFormData.fecha_cita === date;
-  }
-
-  public isDayUnavailable(day: CalendarDay): boolean {
-    return !day.isCurrentMonth || day.isPast || day.availability === 'full';
-  }
-
-  public getDayStyle(day: CalendarDay): { [key: string]: string } | null {
-    if (day.availability === 'none') {
-      return null;
-    }
-
-    const baseColor = day.color ?? this.getStatusColor(day.availability);
-    const background = this.applyOpacity(baseColor, day.availability === 'full' ? 0.25 : 0.15);
-    const border = this.applyOpacity(baseColor, 0.6);
-
-    const style: { [key: string]: string } = {
-      ['background-color']: background,
-      ['border-color']: border
-    };
-
-    if (day.isToday) {
-      style['box-shadow'] = `0 0 0 2px ${this.applyOpacity('#1e88e5', 0.4)}`;
-    }
-
-    return style;
-  }
-
-  public onDateChange(date: string): void {
-    this.createFormData.fecha_cita = date;
-    if (this.createFormData.hora_cita && this.isSlotUnavailable(this.createFormData.hora_cita)) {
-      this.createFormData.hora_cita = '';
-    }
-    this.submitMessage = null;
-    this.buildCalendar();
-  }
-
-  public isSlotUnavailable(slot: string): boolean {
-    if (!this.createFormData.fecha_cita) {
-      return false;
-    }
-
-    const normalizedSlot = this.normalizeTime(slot);
-
-    // Si la cita es para hoy, no permitir seleccionar horarios que ya pasaron
-    if (this.createFormData.fecha_cita === this.today) {
-      const [hours, minutes] = normalizedSlot.split(':').map(Number);
-      const now = new Date();
-      const slotDate = new Date();
-      slotDate.setHours(hours ?? 0, minutes ?? 0, 0, 0);
-
-      if (slotDate.getTime() <= now.getTime()) {
-        return true;
-      }
-    }
-
-    const record = this.getDayRecord(this.createFormData.fecha_cita);
-
-    if (record?.status === 'full') {
-      return true;
-    }
-
-    if (record?.takenSlots.has(normalizedSlot)) {
-      return true;
-    }
-
-    return this.citas.some(cita =>
-      cita.fecha_cita === this.createFormData.fecha_cita &&
-      cita.estatus === 'programada' &&
-      this.normalizeTime(cita.hora_cita) === normalizedSlot
-    );
-  }
-
-  private formatDate(date: Date): string {
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    return `${year}-${month}-${day}`;
-  }
-
-  private formatFullDate(fecha: string): string {
-    const date = this.toDate(fecha);
-    return new Intl.DateTimeFormat('es-MX', {
-      weekday: 'long',
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric'
-    }).format(date);
-  }
-
-  private formatDateTime(date: Date): string {
-    return new Intl.DateTimeFormat('es-MX', {
-      dateStyle: 'medium',
-      timeStyle: 'short'
-    }).format(date);
-  }
-
-  private toDate(fecha: string, hora?: string): Date {
-    const [year, month, day] = fecha.split('-').map(Number);
-    if (hora) {
-      const [hours, minutes] = this.normalizeTime(hora).split(':').map(Number);
-      return new Date(year, (month ?? 1) - 1, day ?? 1, hours ?? 0, minutes ?? 0, 0, 0);
-    }
-    return new Date(year, (month ?? 1) - 1, day ?? 1);
-  }
-
-  private generateTimeSlots(): string[] {
-    const slots: string[] = [];
-    // Horario 8am a 5pm — último slot inicia a las 16:45 (termina a las 17:00)
-    for (let hour = 8; hour < 17; hour++) {
-      ['00', '15', '30', '45'].forEach(minute => {
-        slots.push(`${String(hour).padStart(2, '0')}:${minute}`);
-      });
-    }
-    return slots;
-  }
-
-  public normalizeTime(hora: string): string {
-    const [hours, minutes] = hora.split(':');
-    return `${hours?.padStart(2, '0')}:${(minutes ?? '00').padStart(2, '0')}`;
-  }
-
-  private startOfDay(date: Date): Date {
-    return new Date(date.getFullYear(), date.getMonth(), date.getDate());
-  }
-
-  private startOfMonth(date: Date): Date {
-    return new Date(date.getFullYear(), date.getMonth(), 1);
-  }
-
-  private buildCalendar(): void {
-    const reference = this.startOfMonth(this.currentMonth);
-    const firstDayIndex = reference.getDay();
-    const calendarStart = new Date(reference);
-    calendarStart.setDate(calendarStart.getDate() - firstDayIndex);
-
-    const today = this.startOfDay(new Date());
-    const weeks: CalendarDay[][] = [];
-    let cursor = new Date(calendarStart);
-
-    for (let week = 0; week < 6; week++) {
-      const weekDays: CalendarDay[] = [];
-      for (let dayIndex = 0; dayIndex < 7; dayIndex++) {
-        const date = new Date(cursor);
-        const dateStr = this.formatDate(date);
-        const isCurrentMonth = date.getMonth() === reference.getMonth();
-        const isPast = this.startOfDay(date).getTime() < today.getTime();
-        const isToday = this.startOfDay(date).getTime() === today.getTime();
-        const meta = this.resolveCalendarMetadata(dateStr, isCurrentMonth, isPast);
-
-        weekDays.push({
-          date: dateStr,
-          label: date.getDate(),
-          isCurrentMonth,
-          isToday,
-          isPast,
-          availability: meta.availability,
-          color: meta.color,
-          labelText: meta.label
-        });
-
-        cursor.setDate(cursor.getDate() + 1);
-      }
-      weeks.push(weekDays);
-    }
-
-    this.calendarWeeks = weeks;
-  }
-
-  private resolveCalendarMetadata(
-    dateStr: string,
-    isCurrentMonth: boolean,
-    isPast: boolean
-  ): { availability: CalendarAvailability; color: string; label: string | null } {
-    if (!isCurrentMonth || isPast) {
-      return {
-        availability: 'none',
-        color: this.getStatusColor('none'),
-        label: null
-      };
-    }
-
-    // Domingos: día no laborable
-    const date = this.toDate(dateStr);
-    if (date.getDay() === 0) {
-      return {
-        availability: 'full',
-        color: '#ef5350',
-        label: 'Cerrado'
-      };
-    }
-
-    const record = this.getDayRecord(dateStr);
-    if (!record) {
-      return {
-        availability: 'available',
-        color: this.getStatusColor('available'),
-        label: null
-      };
-    }
-
-    const availability = record.status;
-    const color = record.color ?? this.getStatusColor(availability);
-
-    return {
-      availability,
-      color,
-      label: record.label ?? null
-    };
-  }
-
-  private getDayRecord(date: string): DayAvailabilityRecord | undefined {
-    return this.availabilityMap.get(date);
-  }
-
-  private getStatusColor(status: CalendarAvailability): string {
-    switch (status) {
-      case 'available':
-        return '#1e88e5';
-      case 'partial':
-        return '#ffb300';
-      case 'full':
-        return '#ef5350';
-      default:
-        return '#cfd8dc';
-    }
-  }
-
-  private applyOpacity(hexColor: string, alpha: number): string {
-    const rgb = this.hexToRgb(hexColor);
-    if (!rgb) {
-      return hexColor;
-    }
-
-    const { r, g, b } = rgb;
-    return `rgba(${r}, ${g}, ${b}, ${alpha})`;
-  }
-
-  private hexToRgb(hexColor: string): { r: number; g: number; b: number } | null {
-    let hex = hexColor.replace('#', '');
-    if (hex.length === 3) {
-      hex = hex.split('').map(char => char + char).join('');
-    }
-
-    if (hex.length !== 6) {
-      return null;
-    }
-
-    const bigint = parseInt(hex, 16);
-    return {
-      r: (bigint >> 16) & 255,
-      g: (bigint >> 8) & 255,
-      b: bigint & 255
-    };
-  }
-
-  private loadAvailability(): void {
-    this.isLoadingAvailability = true;
-    const month = this.currentMonth.getMonth() + 1;
-    const year = this.currentMonth.getFullYear();
-
-    this.citaService.getAvailability(month, year)
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(() => of(null)),
-        finalize(() => {
-          this.isLoadingAvailability = false;
-        })
-      )
-      .subscribe(response => {
-        if (response?.days) {
-          this.applyAvailability(response.days);
-        } else {
-          this.availabilityMap = new Map<string, DayAvailabilityRecord>();
-        }
-        this.buildCalendar();
-      });
-  }
-
-  private applyAvailability(days: CitaAvailabilityDay[]): void {
-    const map = new Map<string, DayAvailabilityRecord>();
-
-    days.forEach(day => {
-      const takenSlots = new Set<string>((day.taken_slots || []).map(slot => this.normalizeTime(slot)));
-
-      let status: AvailabilityStatus;
-      if (day.special?.status) {
-        status = day.special.status;
-      } else if (takenSlots.size === 0) {
-        status = 'available';
-      } else if (takenSlots.size >= this.totalSlotsPerDay) {
-        status = 'full';
-      } else {
-        status = 'partial';
-      }
-
-      map.set(day.date, {
-        takenSlots,
-        status,
-        color: day.special?.color ?? undefined,
-        label: day.special?.label ?? null
-      });
-    });
-
-    this.availabilityMap = map;
-  }
-
-  // ===== MÉTODOS DE PRE-EVALUACIÓN IA =====
-
-  loadPreEvaluaciones(): void {
-    this.isLoadingPreEvaluaciones = true;
-    this.preEvaluacionesError = null;
-
-    this.preEvaluacionIAService.getPendientes()
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.preEvaluacionesError = error?.error?.message || 'No se pudieron cargar las pre-evaluaciones';
-          return of({ pendientes: [], total: 0 });
-        }),
-        finalize(() => {
-          this.isLoadingPreEvaluaciones = false;
-        })
-      )
-      .subscribe(response => {
-        this.preEvaluaciones = response.pendientes;
-        this.totalPendientes = response.total;
-      });
-
-    this.preEvaluacionIAService.getPreEvaluaciones()
-      .pipe(takeUntil(this.destroy$), catchError(() => of({ pre_evaluaciones: [] })))
-      .subscribe(response => {
-        this.preEvaluacionesHistorial = (response.pre_evaluaciones ?? [])
-          .filter((p: PreEvaluacion) => p.estatus_validacion !== 'pendiente');
-      });
-  }
-
-  // ===== MÉTODOS DE IA PRIORIDAD =====
-
-  loadPrioridad(): void {
-    this.isLoadingPrioridad = true;
-    this.prioridadError = null;
-
-    this.iaPriorityService.getPendientesPorPrioridad()
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.prioridadError = error?.error?.message || 'No se pudo cargar la clasificación de prioridad.';
-          return of(null);
-        }),
-        finalize(() => { this.isLoadingPrioridad = false; })
-      )
-      .subscribe(res => {
-        this.prioridadResumen = res;
-      });
-  }
-
-  openPreEvaluacionModal(preEvaluacion: PreEvaluacion): void {
-    this.selectedPreEvaluacion = preEvaluacion;
-    this.showPreEvaluacionModal = true;
-    this.comentarioValidacion = '';
-  }
-
-  closePreEvaluacionModal(): void {
-    this.showPreEvaluacionModal = false;
-    this.selectedPreEvaluacion = null;
-    this.comentarioValidacion = '';
-    this.isSubmittingValidacion = false;
-  }
-
-  validarPreEvaluacion(accion: 'validar' | 'descartar'): void {
-    if (!this.selectedPreEvaluacion) return;
-
-    this.isSubmittingValidacion = true;
-
-    this.preEvaluacionIAService.validarPreEvaluacion(
-      this.selectedPreEvaluacion.id,
-      accion,
-      this.comentarioValidacion
-    )
-    .pipe(
-      takeUntil(this.destroy$),
-      catchError(error => {
-        this.preEvaluacionesError = error?.error?.message || 'Error al validar la pre-evaluación';
-        return of(null);
-      }),
-      finalize(() => {
-        this.isSubmittingValidacion = false;
-      })
-    )
-    .subscribe(response => {
-      if (response) {
-        // Actualizar la lista de pre-evaluaciones
-        this.preEvaluaciones = this.preEvaluaciones.filter(
-          p => p.id !== this.selectedPreEvaluacion?.id
-        );
-        this.totalPendientes--;
-        this.closePreEvaluacionModal();
-      }
-    });
-  }
-
-  // ===== PERFIL MÉDICO ALUMNO =====
-
-  openPerfilAlumno(alumnoId: number): void {
-    this.showPerfilAlumnoModal = true;
-    this.perfilAlumnoData = null;
-    this.historialAlumno = [];
-    this.historialAlumnoExpandido = null;
-    this.isLoadingPerfilAlumno = true;
-
-    this.perfilMedicoService.getPerfilAlumno(alumnoId)
-      .pipe(takeUntil(this.destroy$), catchError(() => of(null)), finalize(() => { this.isLoadingPerfilAlumno = false; }))
-      .subscribe(res => { if (res) this.perfilAlumnoData = res.perfil; });
-
-    this.perfilMedicoService.getHistorialAlumno(alumnoId)
-      .pipe(takeUntil(this.destroy$), catchError(() => of(null)))
-      .subscribe(res => { if (res) this.historialAlumno = res.historial; });
-  }
-
-  closePerfilAlumno(): void {
-    this.showPerfilAlumnoModal = false;
-    this.perfilAlumnoData = null;
-    this.historialAlumno = [];
-  }
-
-  toggleHistorialAlumnoItem(id: number): void {
-    this.historialAlumnoExpandido = this.historialAlumnoExpandido === id ? null : id;
-  }
-
-  getEstatusClass(estatus: string): string {
-    switch (estatus) {
-      case 'pendiente': return 'estatus-pendiente';
-      case 'validado': return 'estatus-validado';
-      case 'descartado': return 'estatus-descartado';
-      default: return '';
-    }
-  }
-
-  getConfianzaClass(confianza: number): string {
-    if (confianza >= 0.8) return 'confianza-alta';
-    if (confianza >= 0.6) return 'confianza-media';
-    if (confianza >= 0.4) return 'confianza-baja';
-    return 'confianza-muy-baja';
-  }
-
-  // ===== ESTADÍSTICAS =====
-
-  loadEstadisticas(): void {
-    this.isLoadingEstadisticas = true;
-    this.estadisticasError = null;
-    this.chartsRendered = false;
-
-    this.estadisticasService.getEstadisticas()
-      .pipe(
-        takeUntil(this.destroy$),
-        catchError(error => {
-          this.estadisticasError = error?.error?.message || 'No se pudieron cargar las estadísticas.';
-          return of(null);
-        }),
-        finalize(() => { this.isLoadingEstadisticas = false; })
-      )
-      .subscribe(data => {
-        this.estadisticas = data;
-      });
-  }
-
-  private renderCharts(): void {
-    if (!this.estadisticas) return;
-
-    const meses = this.estadisticas.citas_por_mes;
-
-    // Destruir instancias anteriores
-    this.barChart?.destroy();
-    this.doughnutChart?.destroy();
-
-    // Gráfica de barras: citas por mes
-    this.barChart = new Chart(this.barCanvas.nativeElement, {
-      type: 'bar',
-      data: {
-        labels: meses.map(m => m.label),
-        datasets: [
-          {
-            label: 'Atendidas',
-            data: meses.map(m => m.atendidas),
-            backgroundColor: 'rgba(76, 175, 80, 0.7)',
-            borderColor: '#388E3C',
-            borderWidth: 1
-          },
-          {
-            label: 'Canceladas',
-            data: meses.map(m => m.canceladas),
-            backgroundColor: 'rgba(239, 83, 80, 0.7)',
-            borderColor: '#C62828',
-            borderWidth: 1
-          },
-          {
-            label: 'No asistió',
-            data: meses.map(m => m.no_asistio),
-            backgroundColor: 'rgba(255, 179, 0, 0.7)',
-            borderColor: '#F57F17',
-            borderWidth: 1
-          }
-        ]
-      },
-      options: {
-        responsive: true,
-        plugins: {
-          legend: { position: 'top' },
-          title: { display: false }
-        },
-        scales: {
-          y: { beginAtZero: true, ticks: { stepSize: 1 } }
-        }
-      }
-    });
-
-    // Gráfica donut: distribución total de estados
-    const r = this.estadisticas.resumen_estados;
-    this.doughnutChart = new Chart(this.doughnutCanvas.nativeElement, {
-      type: 'doughnut',
-      data: {
-        labels: ['Atendidas', 'Canceladas', 'No asistió', 'Programadas'],
-        datasets: [{
-          data: [r.atendida, r.cancelada, r.no_asistio, r.programada],
-          backgroundColor: [
-            'rgba(76, 175, 80, 0.8)',
-            'rgba(239, 83, 80, 0.8)',
-            'rgba(255, 179, 0, 0.8)',
-            'rgba(33, 150, 243, 0.8)'
-          ],
-          borderColor: ['#388E3C', '#C62828', '#F57F17', '#1565C0'],
-          borderWidth: 2
-        }]
-      },
-      options: {
-        responsive: true,
-        plugins: {
-          legend: { position: 'bottom' }
-        }
-      }
-    });
   }
 }
